requirements.txt
asgiref==3.3.4
boto3==1.17.95
botocore==1.20.95
certifi==2021.5.30
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.2
django-cors-headers==3.7.0
django-import-export==2.5.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
Markdown==3.3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
psycopg2==2.8.6
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.5
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
